Move implementation-independent verification methods into trait
Both verify_mirror_property and verify_unitig_length_is_zero do not check implementation specific problems, but verify properties that not all bigraphs must fulfill.
Therefore, they are valid for all bigraph implementations and can be in the trait.
